% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw04phayleymanke10-package.R
\docType{package}
\name{hw04phayleymanke10-package}
\alias{hw04phayleymanke10}
\alias{hw04phayleymanke10-package}
\title{hw04phayleymanke10: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Hayley Manke \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
